Changed param format
git-svn-id: https://outlook-client.forge.funambol.org/svn/client-sdk/branches/java-sdk@3278 3966be30-755a-0410-815e-dd79f492ac3c

+++ us7638/client/src/main/java/com/funambol/client/sapi/SapiHandler.java
@@ -125,7 +125,7 @@ public class SapiHandler {
 
 
     public JSONObject query(String name, String action, Vector params, 
-            Hashtable headers, InputStream requestIs, int contentLength)
+            Hashtable headers, InputStream requestIs, long contentLength)
             throws IOException, JSONException {
         String url = createUrl(name, action, params);
         HttpConnectionAdapter conn;
